Physics: Allow deferred shape loading
When adding bodies to the world of the bullet engine, the body must have a valid
shape and we cannot modify the shape. However, we need the ability to change
shapes during runtime. Therefore, we allow to add phys_body's to a phys_world
without actually having a shape. But whe do not add the object to the real world
unless a shape is loaded.
Users shouldn't notice this and this is handled in the physics abstraction.
However, users will notice that objects without a shape have no effects on the
physics world. But this is normal and expected.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>