World: Add world tree
The world tree is a tree of objects. Each object is assigned a shape and physics
body. Both are by default empty and have no effect but you may modify them to
your needs.
World objects always own their childs, that is, if you unlink an object, you
also unlink all its childs. This allows to move complex objects or even scenes
around in world space.

The world object is the root object. If an object should be drawn on the
screen, it must be linked to a world. The world also takes care for physics so
each object is always added to the physics world, however, it does not have any
effect unless a physics shape is added.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>