Scripts: Add SFML key converter
Add new script that helps converting SFML keys to match airhockey keys. This
allows us to easily create an enum that maps airhockey keys to SFML keys and we
do not need any conversion.
This script can be used when the SFML header adds or removes keys and we need to
update our enum.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>