Physics: Allow constant forces on single bodies
Add function to apply constant forces on single bodies. We do not use
applyCentralForce() as this is only for one physics world step. Instead we add
the force to the gravity to allow constant forces.
Its important to do this after adding an object to the world, because adding the
object to the world reset gravity to the default.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>