Rewrite E3D buffers
Instead of using one buffer object for all three components (vertex, color,
normals), we now have a generic e3d_vbo implementation which supports any kind
of data and can be used independently from each other.
We now have a slightly bigger memory consumption, but we are also able to share
vertex and color buffers between multiple objects while having an own normal
buffer for each object.
This will be heavily used for circles, cylinders and more.

This also required to rewrite the e3d_primitive and since we were rewriting it
anyway, we also adjusted the e3d_shape API.

The whole config loader has been rewritten to work with the new buffers. It was
also split into multiple source files since in the future this may extend
heavily and debug support may be added.

We needed this rewrite to reduce memory consumption for objects that share
vertex and color buffer but not the normals buffer. Although the new buffer
objects require more memory than the old ones, we will get a better memory usage
overall as we can share the objects.
We currently do not support geometry shaders as Shader Model 2.0 does not
support it and we want the engine to run on old graphic cards, too. However,
with geometry shaders we could compute normals in the graphic card and hence
would not need any normal buffers in the application. This may be supported in
the future to reduce memory consumption even more. Geometry shaders will be
optional, though.

Data objects were also adjusted to the new config file format. Next step will be
adding other objects than raw primitives. This will allows us to reduce the
config files to a minimum and make them more readable.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>